+++ api/src/resources/dataset.rs
@@ -301,7 +301,7 @@ pub(crate) struct UpdateResponse {
 #[cfg(test)]
 mod tests {
     use crate::resources::comment::{
-        CommentTimestampFilter, UserPropertiesFilter, UserPropertyFilterKind,
+        CommentTimestampFilter, PropertyFilterKind, UserPropertiesFilter,
     };
 
     use super::*;
@@ -431,7 +431,7 @@ mod tests {
     pub fn test_serialize_user_properties_request_params() {
         let user_property_filter = UserPropertiesFilter(HashMap::from([(
             "string:Generation Tag".to_string(),
-            UserPropertyFilterKind::OneOf(vec![PropertyValue::String(
+            PropertyFilterKind::OneOf(vec![PropertyValue::String(
                 "72b01fe7-ef2e-481e-934d-bc2fe0ca9b06".to_string(),
             )]),
         )]));